Fix gitlab build (#5907)
## Summary of changes

- Fixes the gitlab build that I broke by merging #5744
- Runs the build on every push

## Reason for change

I broke the build on master. It's happened one too many times, so
changed this to _always_ do the build (i.e. in builds too). Seeing as
that's what we're doing for one pipeline now anyway, doesn't make much
sense to _not_ do it for the main build too IMO.

## Implementation details

- Removed the `only`
- Fixed the folder conflict in the gitlab job

Yes, we could probably do something tidier, no I couldn't be bothered
right now 😅

## Test coverage

This is the test - if it builds, we should be good 🤞 

## Other details

Docker on Windows in GitLab is so slow 😅