[serverless] No-op AWS Lambda integration on missing API Key (#5900)
## Summary of changes

Verifies that an API key is set in order to call the Lambda Extension.

## Reason for change

It would not crash the customers, but it logs errors that should be
avoided and handled better by not doing the operation.

## Implementation details

Just created an API key getter in order to then check if the API is
missing.

## Test coverage

Tested manually on AWS Lambda.
<img width="1486" alt="Screenshot 2024-08-14 at 9 00 16 PM"
src="https://github.com/user-attachments/assets/7a993627-3b68-4ef1-9960-3d69890a8f56">

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->