Pin `StyleCop.Analzyers` to latest pre-release (#5908)
## Summary of changes

:sorry:

Pins `StyleCop.Analyzers` to the latest prerelease instead of a floating
version.
Reverts ignoring `StyleCop.Analyzers` in Dependabot because that did
nothing.

## Reason for change

Dependabot still struggling with ` error : '[1.2.0-beta.*]' is not a
valid version string`

## Implementation details

Updated both locations of `StyleCop.Analzyer` to the latest pre-release
Dependabot will attempt to update this package again

## Test coverage

- Nothing

## Other details
If this doesn't fix it I'll find a better way to update/test this
without having to PR and merge to run dependabot.
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Quinna Halim <quinna.halim@datadoghq.com>