[Tracer] Skip inserting the startup hook into methods in the type Costura.AssemblyLoader (#5910)
For automatic instrumentation, the tracer inserts a startup hook inside the first method that we deem as a "valid" startup hook callsite. This change adds logic so that methods inside the `Costura.AssemblyLoader` type are not "valid" startup hook callsites, so that the insertion of the startup hook can be delayed to a later method call.