Use a native logger for critical failures in the loader (#5929)
## Summary of changes

Wrap the call to the managed loader in a try/catch, and use a native
logger to log the exception if any.

## Reason for change

There's already a try/catch in the managed loader, but in some
situations we can fail before even getting there (if jitting the method
failed, for instance because of a trimmed runtime). Also, the managed
loader logger can fail.

## Implementation details

This PR adds a try/catch in the code emitted to load the managed loader.
In the catch clause, we call a native logging function, with a direct
unmanaged `calli` call to its address.

Because we have now this outer catch block, I removed the `IsInSsi`
logic which isn't needed anymore.