Revert "[serverless] No-op AWS Lambda integration on missing API Key" (#5941)
## Summary of changes

This reverts commit 8e63384717cc1a2608815c661c13c65ed94ba965.

## Reason for change

It doesn't account for `DD_API_KEY_SECRET_ARN` or `DD_KMS_API_KEY` which
is considered a regression.

## Implementation details

n/a

## Test coverage

Confirmed that it stops working for Lambdas with that configuration

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->