Remove dependency of download-single-step-artifacts on build (#5945)
## Summary of changes
`download-single-step-artifacts` doesn't depend on any other jobs

## Reason for change
With the current pipeline, the download job doesn't start until the
`build` job finishes. The `build` job, despite the genericness of the
name, is only responsible for building windows artifacts. This can
sometimes take dozens of minutes delaying the rest of the pipeline for
no reason