Prevent the native loader from being unloaded while sending telemetry (#5944 => V2) (#5957)
This is a backport of #5944 to v2.

## Summary of changes

Prevent the native loader from being unloaded while sending telemetry.

## Reason for change

We send telemetry when we decide not to instrument a process (for
instance because of an EOL runtime). To send the telemetry, we spawn the
telemetry forwarder. This is done in a background thread to avoid
blocking the startup of the process. However, .NET unloads the profiler
when the `CorProfilerInfo::Initialize` method returns, even if the
background thread is still running, causing a segfault.

## Implementation details

We increment the reference count of the module to prevent it from being
unloaded by the .NET runtime.

Two different implementations:
- On Windows, we use `GetModuleHandleEx` to increment the reference
count. Then, when we're done sending the telemetry, we use
`FreeLibraryAndExitThread` to safely unload the module.
- On Linux, we use `dlopen` to increment the reference count.
Unfortunately there is no safe way to unload the module from within
itself, so we keep it in memory.

If for some reason we failed to increment the reference count, we give
up on sending telemetry.

## Test coverage


`OnEolFrameworkInSsi_WhenForwarderPathExists_CallsForwarderWithExpectedTelemetry`
segfaults on 3.0 without this change.