[ASM] RASP shell injection vulnerability (#5871)
## Summary of changes

This PR contains the implementation of the RASP shell injection
vulnerability. This vulnerability is defined in [this
RFC](https://docs.google.com/document/d/1gCXU3LvTH9en3Bww0AC2coSJWz1m7HcavZjvMLuDCWg/edit#heading=h.giijrtyn1fdx)

This vulnerability is only launched when a process is executed from a
shell. That includes both processes with the useShellExecute flag set to
true in the ProcessStartInfo class.

It can also be a case where a command does not have that flag but calls
a shell executable, but this case is not yet supported but it is
expected to be supported in next WAF releases.

## Reason for change

Is required for RASP during this quarter.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->