Add support for alpine on arm64  (#5933)
## Summary of changes

Adds support for the tracer on alpine on ARM64

## Reason for change

We currently support x64 on glibc/alpine but only glibc on arm64. This
closes that matrix which makes it easier to understand

## Implementation details

There's several layers, but at a high level
- Anywhere where we detect platform, make sure to include
`linux-musl-arm64` as well as `linux-arm64`
- Build the tracer + profiler on alpine arm64

There were several difficulties
- The ruby-based fpm bundler we use doesn't work on alpine arm64, hence
#5905
- I couldn't get our existing alpine.build.dockerfile to compile on
arm64 (where we currently compile llvm/clang16 on alpine14).
- To work around it, created a separate alpine.build.arm64.dockerfile
image which uses alpine:18 (the first version that includes clang16) as
the base.
- Obviously this gives us different requirements in arm64, but I think
that's fine because a) this is a new feature b) we already have
differences on the glibc requirements for arm64 c) We only support .NET
5+ anyway on arm64 (.NET 6 really given .NET 5 is mostly gone) and they
ship alpine:18 images for .NET 6+
- I created a multi arch base image from the two base images ([as some
bloke with a blog describes
here](https://andrewlock.net/combining-multiple-docker-images-into-a-multi-arch-image/))
- I had difficulty compiling the profiler on arm64 initially (we require
it for some crash tracking functionality), as we were getting `undefined
reference to `_Uaarch64_get_accessors_int'` when trying to compile
libunwind. @gleocadie tracked it down to [this
issue](https://github.com/libunwind/libunwind/issues/788), and fixed it
by just not building the libunwind tests 🙇‍♂️
- The sqlite IAST instrumentation tests on alpine arm64 were failing in
.NET 5. I _think_ that's because the sqlite package we're using there
doesn't support it (which makes sense) so just disabled the tests for
that scenario. They pass in .NET 6+ so I'm not worried about it.
- Removed the "noop" smoke tests stage (there's nothing to run in it
currently)
- We can't even install .NET Core 2.1 using the dotnet install script on
alpine arm64, so bailing out. We can probably bail out of more too, but
that's not _necessary_ so haven't done it yet.

There's still one outstanding issue I can't get my head around, on
_debian x64_ (i.e. not something that should have changed) The
`OnEolFrameworkInSsi_WhenForwarderPathExists_CallsForwarderWithExpectedTelemetry`
tests is failing by crashing. It's bizarre, I can't figure it out, if
anyone has any ideas, please let me know 😅

## Test coverage

I've run full installer + full TFM tests
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=162869&view=results)
🤞

## Other details

Stacked on 
- https://github.com/DataDog/dd-trace-dotnet/pull/5770
- https://github.com/DataDog/dd-trace-dotnet/pull/5905

as they both change key parts of the build + will mean we need to
rebuild the VMs

Fixes #3850

---------

Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>