[ASM] Avoid reporting unknown matcher WAF errors (#5955)
## Summary of changes

The unknown matcher error happens when a new operator is added and it is
not supported by the WAF, so sending this particular error is causing a
lot of noise because old versions of the WAF will return that error that
actually is supported in newer versions. We already write warnings for
this situation, but we are also reporting errors that are not actually
errors, Which is generating a lot of noise. In particular, we return
errors in the remote config operations and we add error tags to the
span.

This PR corrects this.

## Reason for change

It was reported and requested.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->