[ASM] Avoid unhandled HttpRequestValidationExceptions (#5943)
## Summary of changes

We are logging some not handled HttpRequestValidationException errors

[In the past](https://github.com/DataDog/dd-trace-dotnet/pull/5221), we
protected the reading of the query string to avoid such issues, but now
they seem to be caused by the values of headers and cookies.

This error occurs in .Net framework. If the method request.ValidateInput
is called, the request will launch an exception if the cookies, url or
header values are dangerous.

In the past, we created a class called QueryStringValidator. In this PR,
the features of this class are extended to cookies and headers and urls.
The class has been renamed to RequestDataValidator.

## Reason for change

## Implementation details

## Test coverage

Some unit tests have been added.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->