[Profiler] Support Single Step Instrumentation deployment and activation (#5240)
## Summary of changes
Add IProfilerTelemetry definition and implementation to be able to send metrics related to SSI deployment and impacts on the generated or not profiles based on heuristics (first span created and process lifetime > 30 seconds).
The profiler is now able to start with a delay based on the heuristics.

## Reason for change
Get statistics about heuristics effects when deployed with SSI and support delayed start of the profiler based on heuristics.

## Implementation details
Major refactoring to allow delay start of the profiler.

## Test coverage
Corresponding tests have been added both for the configuration and the runtime behavior depending on new "auto" value and "profiler" in env vars.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>
Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>