DSM Full propagation mode for SQL Server (#5859)
## Summary of changes

Adding support for `Full` propagation mode for SQL Server when using
DSM.
DSM has 3 modes: disabled, service and full.
Service means that we inject info about the service who sent the query
Full means Service + trace ID and span ID.

## Reason for change

We used to forbid Full mode for sql server because it uses a cache of
query text -> execution plan, and injecting with comments like we did
before would break the cache. (injecting service is OK because it
doesn't change much, so the cache is still OK)
The solution that was found was to use the `context_info` to set the
trace/span ID. It's independant from the query itself, and sets a
"context" for the duration of the session that is visible to DBM when
they sample a "slice" of the DB.
The drawback ofc is that it's an extra query (albeit a fast one) that we
have to run before every "regular" query. It can be slow mostly if the
network is slow. For this reason, we want to cover it with a span to
clearly represent the time being spent there, so that clients can
identify where the time is spent.

## Implementation details

## Test coverage

I added this scenario to the integration test, but since the query is
parameterized, I don't think I can assert easily on the value that is
set.

## Other details

This is the dotnet equivalent of
https://github.com/DataDog/dd-trace-java/pull/7186

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>