Make sure we run all the TFMs on master builds (#5990)
## Summary of changes

Update `RequiresThoroughTesting()` to return `true` when we're running
on master or release/2.x

## Reason for change

We should be doing thorough testing on the main branches. I thought we
were. I don't quite understand how we're _not_

## Implementation details

Check to see if we're _on_ the main branch for the verison, and if so,
do the thorough testing

## Test coverage

Can't really test this properly until we merge it...

## Other details

I sure hope this doesn't reveal more issues :awkward: