[Crashtracking] Fix the handling of COMPlus_DbgMiniDumpName (#5980)
## Summary of changes

To distinguish the case when createdump is called because of a crash vs
a call from `dotnet-dump`, we store a dummy value in
`COMPlus_DbgMiniDumpName`:
https://github.com/DataDog/dd-trace-dotnet/pull/5852

However, this becomes a problem when the process spawns a child process,
as it inherits the dummy `COMPlus_DbgMiniDumpName` and we lose the
original value.

To get around that, this PR saves the original value in
`DD_INTERNAL_CRASHTRACKING_MINIDUMPNAME` so the children can retrieve
it.

## Reason for change

This prevents the crash dump to be generated when launching the app from
a bash script when `LD_PRELOAD` is set globally.

## Implementation details

We discovered that bash provides its own version of the
`getenv`/`setenv` functions, which leads to some unexpected behavior. To
get around that, we use `dlsym` to fetch the original libc functions.

## Test coverage

Added a test with a sample app launched from a bash script.