Add proper error checking around GetModuleMetadata (#5985 => v2) (#5992)
## Summary of changes

Explicitly check for `S_OK` when calling `GetModuleMetadata`.

## Reason for change

`GetModuleMetadata` can return `S_FALSE` (one known case is if the
module is a resource, but there might be other). We've been using the
`SUCCEEDED`/`FAILED` macros to check the error code, but `S_FALSE` is
not considered as an error so we would mistakenly assume that the call
succeeded.

## Implementation details

Note that I only updated the error checks. There are a couple of places
where we call `GetModuleMetadata` without checking the error code, I
haven't touched those.

For `GetModuleInterfaces` in IAST, I also moved the call to `Release` to
only do it if the call succeeded.