Fix installer tests (#5994 => main)
- Fix nuget/dd-trace installer tests for the future

The code we were using to install the "local" builds of the NuGet
packages added the local source. However the dotnet restore was looking in both the local and nuget.org sources, and installing the highest version it found. That worked fine until we released 3.2.0 publicly and expect to install 2.59.0 of the local build.

Install an explicit version of the tracer. As this version is never in
the public nuget.org source but _is_ in the local source, it uses that.

Having to thread the version through everywhere is kinda horrible, but
is the only thing I could find that works.

The whole approach should be rewritten to be managed by Nuke tbh, if we ever find time