Fix 2.x smoke test and installer tests (#5994)
## Summary of changes

- Fix nuget/dd-trace installer tests in 2.x branch
- Pin system tests in 2.x branch

## Reason for change

The code we were using to install the "local" builds of the NuGet
packages added the local source. However the dotnet restore was looking
in both the local and nuget.org sources, and installing the highest
version it found. That worked fine until we released 3.2.0 publicly and
expect to install 2.59.0 of the local build.

For system-tests, we had to change how things work to support v3
(https://github.com/DataDog/system-tests/pull/2962), but unfortunately
that breaks for v2.

## Implementation details

Install an explicit version of the tracer. As this version is never in
the public nuget.org source but _is_ in the local source, it uses that.

> In https://github.com/DataDog/dd-trace-dotnet/pull/5989 I tried a
different approach but it just didn't work.

For system tests, just pinning to the specific commit prior to the v3 PR
being merged.

Having to thread the version through everywhere is kinda horrible, but
is the only thing I could find that works.

## Test coverage

Running [a full installer run
here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=163526&view=results)

## Other details

The whole approach should be rewritten to be managed by Nuke tbh, if we
ever find time

This PR also contains a backport of 
- https://github.com/DataDog/dd-trace-dotnet/pull/5988
for simplicity

Supersedes
- https://github.com/DataDog/dd-trace-dotnet/pull/5989