Fix installer tests (#5994 => main) (#6004)
## Summary of changes

- Fix nuget/dd-trace installer tests in 2.x branch

## Reason for change

The code we were using to install the "local" builds of the NuGet
packages added the local source. However the dotnet restore was looking
in both the local and nuget.org sources, and installing the highest
version it found. That worked fine until we released 3.2.0 publicly and
expect to install 2.59.0 of the local build.

## Implementation details

Install an explicit version of the tracer. As this version is never in
the public nuget.org source but _is_ in the local source, it uses that.

> In https://github.com/DataDog/dd-trace-dotnet/pull/5989 I tried a
different approach but it just didn't work.

Having to thread the version through everywhere is kinda horrible, but
is the only thing I could find that works.

## Test coverage

Will run a [full installer
test](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=163706&view=results)

## Other details

The whole approach should be rewritten to be managed by Nuke tbh, if we
ever find time

This PR is a "forward port" of 
- https://github.com/DataDog/dd-trace-dotnet/pull/5994

(the issue was exhibited on the 2.x branch, so this is preemptive for
the future)