Fix SSI tests for profiler integration tests (#6016)
## Summary of changes

Fix profiler integration tests in the SSI run

## Reason for change

#5240 changed the behaviour of the profiler in the presence of SSI
variables. We have a scheduled run on master that specifically sets the
variables. This breaks the profiler integration tests that depend on
those variables.

## Implementation details

Add an "environment restorer" attribute, which mirrors the one we use in
the tracing integration tests. Ensures that the SSI tests are working in
their expected environment

## Test coverage

This is the test, but did [a dedicated run
too](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=163728&view=results),
and it works

## Other details

Related to #6014