Fix gitlab build (#6025)
## Summary of changes

Enables longpath support on the gitlab Windows jobs

## Reason for change

We moved the location of the repo in GitLab, which increased the path
length, and we hit long path issues

## Implementation details

Make the runner set `git config --system core.longpaths true` - it's [a
crazy olde time
restriction](https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=powershell),
so realistically everyone should have this anyway

## Test coverage

The build works, so it's fixed!

## Other details
We could shorten the paths to work around it, but it's just another
accident waiting to happen. This is not the first time 🙈

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->