[Tool] update continuous profiler diagnostics (#6014)
## Summary of changes

Fix the diagnostics associated with the expected state of the continuous
profiler

## Reason for change

In #5240, the continuous profiler gained two new "modes" of operation:
delayed enablement based on heuristics, and "monitoring" mode. The tool
should be updated to understand those possibilities + handle it in the
tests

## Implementation details

- Reset the `DD_INJECTION_ENABLED` variable before running the tests -
the presence/absence of the variable changes the expectations, so we
need to account for it
- Explicitly test the various combinations of `DD_INJECTION_ENABLED` and
`DD_PROFILING_ENABLED`.

## Test coverage

Added additional integration and artifact tests

## Other details

This should address most of the issues in the scheduled SSI run. I
suspect the windows integration tests _may_ need additional work, as
currently the profiler library isn't copied to those tests. Will fix in
this PR if that ends up being the case

Supersedes 
- https://github.com/DataDog/dd-trace-dotnet/pull/5991

---------

Co-authored-by: chrisnas <chrisnas@users.noreply.github.com>