[ASM]Fix InvalidOperationException in httpContext.Items (#6017)
## Summary of changes

This PR fixes the InvalidOperationException that is thrown in the ASM
code.
Some logs can be found
[here](https://app.datadoghq.com/logs?query=source%3Adotnet%20%2AAppsec%2A%20%40tracer_version%3A3%2A%20%2AInvalidOperationException%2A%20&agg_m=count&agg_m_source=base&agg_t=count&cols=host%2Cservice&event=AgAAAZHXAWGCGYxJ7AAAAAAAAAAYAAAAAEFaSFhBV0dDQUFBQmh4MUdySmhmTkFBQQAAACQAAAAAMDE5MWQ3MDctMWNiYS00NDgwLWJhODQtZWY5YTk4NmQ4YzFk&fromUser=true&messageDisplay=inline&refresh_mode=sliding&storage=hot&stream_sort=desc&viz=stream&from_ts=1720177260603&to_ts=1720782060603&live=true).

This error can happen only in netcore versions of the framework. When we
access httpcontext.Items, the exception is thrown if the key is not
found. httpcontext.Items is defined as a IDictionary<object, object>.
Usually, it will be a Microsoft.AspNetCore.Http.ItemsDictionary, which
does not throw an exception when trying to retrieve a key that is not
stored, but other custom implementations such as Dictionary<object,
object> will throw it. It seems that in one customer, we are receiving a
context in which Items is a Dictionary.

This might be due to custom middlewares, third party extensions, use of
custom http contexts, etc.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->