[IAST] Add Stack trace to vuln location (#5997)
## Summary of changes
Add full stack trace to vulnerability location following this
[RFC](https://docs.google.com/document/d/1ga7yCKq2htgcwgQsInYZKktV0hNlv4drY9XzSxT-o5U/edit)

## Reason for change
A stack trace of where the vulnerability is located can help the client
to locate it in order to fix it.

## Implementation details
The stacks are added in the MetaStruct `_dd.stack` tag under the
`vulnerability` section. A stack Id has been added to the vulnerability
location. This Id matches the stack Id on the MetaStruct dictionary.

## Test coverage
Added integration tests

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->