[ASM] Attacker fingerprint (#5982)
## Summary of changes

Fingerprinting is a technique used to identify and track users through
the use of available data which, when combined through a certain set of
algorithms, can provide a unique fingerprint for said user.
Fingerprinting can be performed on many contexts with different data
sets, such as the browser, which can provide the algorithm with specific
data about the user’s software and hardware stack, or the server, which
typically provides data at the different levels of the network stack.

This PR contains the implementation of the attacker fingerprint feature
described in [this
RFC](https://docs.google.com/document/d/1DivOa9XsCggmZVzMI57vyxH2_EBJ0-qqIkRHm_sEvSs/edit?pli=1).

## Reason for change

## Implementation details

There are two small issues detected that seem related to the WAF:
If we don't send the request body, no endpoint fingerprint
(_dd.appsec.fp.http.endpoint) is generated.
The agent header fingerprint is not generated if we send a value in a
dictionary instead of a regular string.

These issues will be discussed with the libdwaf team.

## Test coverage

Some unit tests have been added. 

Since this feature will be enabled by default and, in order to cover
different situations while not impacting the CI performance, the ASM
integration tests have been modified to include the fingerprint in the
snapshots.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->