Bump timeit to 0.1.21 (#6059)
## Summary of changes

Bumping timeit to version 0.1.21, also adding a script to timeit to
remove the logs generated by the tracer before running each scenario.

The PR also bumps the number of iteration to check if we can reduce the
error of the metrics.

The PR also adds a new configuration file for net8.0

## Reason for change

On the benchmark dashboard we can see some unexplained bumps on the same
commit (increasing duration on startup and thread count) when using the
tracer. We are trying to stabilize the values by removing the logs to
avoid the rollup logics from the logs on startup.

## Implementation details

The new version of timeit has support for executing commands before any
timeit event. We added a new remove_logs.cmd script to run it before
each scenario start and after each execution the target process.

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->